More sweeping changes on error control, need to consolidate PrivateKey type (from ssh_keys, rustls, our custom one...). Also the testsuite is too based on filesystem. [ci skip]
