+++ src/lib.rs
@@ -32,6 +32,7 @@ pub mod keys;
 pub mod error;
 
 use bytes::Bytes;
+use decrypter::DecrypterStream;
 use io::Cursor;
 use keys::SessionKeys;
 use rand::{SeedableRng, RngCore};
@@ -256,14 +257,14 @@ pub fn decrypt(
 
     // Todo crypt4gh-rust body_decrypt_parts does not work properly, so just apply edit list here.
     body_decrypt(read_buf, session_keys.as_slice(), &mut write_info, 0)
-      .map_err(|err| Crypt4GHError(err.to_string()))?;
+      .map_err(|err| Crypt4GHError::UnableToDecryptBlock(read_buf, err.to_string()))?;
     let mut decrypted_bytes: Bytes = write_buf.into_inner().into();
     let mut edited_bytes = Bytes::new();
 
     let edits = DecrypterStream::<()>::create_internal_edit_list(edit_list_packet)
       .unwrap_or(vec![(false, decrypted_bytes.len() as u64)]);
     if edits.iter().map(|(_, edit)| edit).sum::<u64>() > decrypted_bytes.len() as u64 {
-      return Err(Crypt4GHError(
+      return Err(Crypt4GHError::UnableToDecryptBlock(edits,
         "invalid edit lists for the decrypted data block".to_string(),
       ));
     }
